Declare onRequestPermissionResult instead of onRequestPermissionsResult
As noted in the comment, the former is deprecated in favour of the
latter, but the latter is never called. See also:

- https://github.com/apache/cordova-android/issues/1388
- https://github.com/apache/cordova-android/issues/1393

I hope the comment will help us fix the problem quickly when the Cordova
folks decide to remove the deprecated method. (And hopefully make sure
that the new method is called.)

Addresses: https://trello.com/c/zsmadleg/2549-mind-ease-notifications-probably-broken-on-android