Complete scripts and README
The README is something to get people started.

The scripts are a uniform way to do things that were previously done by
hand (install-deps) or handled worsely using env.sh.

- Inspiration for putting together a classpath without fancy Java
  machinery: https://tonsky.me/blog/python-build/
- Inspiration for hard-coding configuration: https://dwm.suckless.org/
  via https://ericnormand.substack.com/p/complexity-begets-complexity.

Yes, it's Clojure, executed using Babashka. What are the alternatives?
Ruby – no. Bash – no. Python – yeah, but also a language no-one else
on the team uses day-to-day. JavaScript – yeah, but would need to
research libraries to do the things Babashka has built-in.

Addresses: https://trello.com/c/4PqplE1e/2550-mind-ease-js-tests-are-failing-and-other-small-issues